+++ modules/standard-client-saml/main.tf
@@ -35,17 +35,6 @@ resource "keycloak_saml_client_default_scopes" "idp_scopes" {
   default_scopes = [for v in var.idps : "${v}-saml"]
 }
 
-resource "keycloak_generic_client_protocol_mapper" "audience_mapper" {
-  realm_id        = var.realm_id
-  client_id       = module.standard_saml_client.id
-  name            = "audience"
-  protocol        = "saml"
-  protocol_mapper = "saml-audience-mapper"
-  config = {
-    "included.client.audience" : var.client_id
-  }
-}
-
 resource "keycloak_generic_client_protocol_mapper" "additional_client_roles" {
   count           = var.additional_role_attribute != "" ? 1 : 0
   realm_id        = var.realm_id